Daddy's WordCloud/requirements.txt
requests
beautifulsoup4
jieba
wordcloud
Scrapy
numpy
Pillow
snownlp
mysql-connector-python